+++ stories/Shape/component/card-node.tsx
@@ -17,9 +17,9 @@ function fetchData() {
       nodeType: 'a',
       collapse: true,
       panels: [
-        {title: '成功率', value: '11%' }, 
-        {title: '耗时',value: '111'}, 
-        {title:'错误数', value: '111'}
+        { title: '成功率', value: '11%' },
+        { title: '耗时', value: '111' },
+        { title: '错误数', value: '111' }
       ],
     },
     {
@@ -30,9 +30,9 @@ function fetchData() {
       nodeType: 'b',
       collapse: true,
       panels: [
-        {title: '成功率', value: '11%' }, 
-        {title: '耗时',value: '111'}, 
-        {title:'错误数', value: '111'}
+        { title: '成功率', value: '11%' },
+        { title: '耗时', value: '111' },
+        { title: '错误数', value: '111' }
       ],
     }
   ]
@@ -62,7 +62,7 @@ const EXPAND_ICON = function EXPAND_ICON(x, y, r) {
 
 const ICON_MAP = {
   a: 'https://gw.alipayobjects.com/mdn/rms_8fd2eb/afts/img/A*0HC-SawWYUoAAAAAAAAAAABkARQnAQ',
-  b: 'https://gw.alipayobjects.com/mdn/rms_8fd2eb/afts/img/A*sxK0RJ1UhNkAAAAAAAAAAABkARQnAQ', 
+  b: 'https://gw.alipayobjects.com/mdn/rms_8fd2eb/afts/img/A*sxK0RJ1UhNkAAAAAAAAAAABkARQnAQ',
 }
 
 G6.registerNode(
@@ -83,8 +83,8 @@ G6.registerNode(
         name: 'main-box',
         draggable: true,
       });
-      
-      group.addShape('rect',  {
+
+      group.addShape('rect', {
         attrs: {
           x: 0,
           y: 0,
@@ -96,7 +96,7 @@ G6.registerNode(
         name: 'title-box',
         draggable: true,
       });
-      
+
       // 左侧图标
       group.addShape('image', {
         attrs: {
@@ -111,7 +111,7 @@ G6.registerNode(
       });
 
       // 标题
-      group.addShape('text',  {
+      group.addShape('text', {
         attrs: {
           textBaseline: 'top',
           y: 2,
@@ -124,7 +124,7 @@ G6.registerNode(
       });
 
       if (cfg.nodeLevel > 0) {
-         group.addShape('marker', {
+        group.addShape('marker', {
           attrs: {
             x: 184,
             y: 30,
@@ -137,11 +137,11 @@ G6.registerNode(
           name: 'collapse-icon',
         });
       }
-     
+
       // 卡片节点中的内容列表
-      cfg.panels.forEach((item,  index) => {
+      cfg.panels.forEach((item, index) => {
         // 名称
-        group.addShape('text',  {
+        group.addShape('text', {
           attrs: {
             textBaseline: 'top',
             y: 25,
@@ -154,7 +154,7 @@ G6.registerNode(
         });
 
         // 值
-        group.addShape('text',  {
+        group.addShape('text', {
           attrs: {
             textBaseline: 'top',
             y: 42,
@@ -165,7 +165,7 @@ G6.registerNode(
           },
           name: `index-title-${index}`
         });
-       
+
       });
       return shape;
     },
@@ -180,9 +180,9 @@ const data = {
   id: 'root',
   nodeLevel: 2,
   panels: [
-    {title: '成功率', value: '11%' }, 
-    {title: '耗时',value: '111'}, 
-    {title:'错误数', value: '111'}
+    { title: '成功率', value: '11%' },
+    { title: '耗时', value: '111' },
+    { title: '错误数', value: '111' }
   ],
   children: [
     {
@@ -192,9 +192,9 @@ const data = {
       id: 'node1',
       nodeLevel: 2,
       panels: [
-        {title: '成功率', value: '11%' }, 
-        {title: '耗时',value: '111'}, 
-        {title:'错误数', value: '111'}
+        { title: '成功率', value: '11%' },
+        { title: '耗时', value: '111' },
+        { title: '错误数', value: '111' }
       ],
       collapse: true
     },
@@ -205,9 +205,9 @@ const data = {
       id: 'node2',
       nodeLevel: 0,
       panels: [
-        {title: '成功率', value: '11%' }, 
-        {title: '耗时',value: '111'}, 
-        {title:'错误数', value: '111'}
+        { title: '成功率', value: '11%' },
+        { title: '耗时', value: '111' },
+        { title: '错误数', value: '111' }
       ],
     },
     {
@@ -217,9 +217,9 @@ const data = {
       id: 'node3',
       nodeLevel: 3,
       panels: [
-        {title: '成功率', value: '11%' }, 
-        {title: '耗时',value: '111'}, 
-        {title:'错误数', value: '111'}
+        { title: '成功率', value: '11%' },
+        { title: '耗时', value: '111' },
+        { title: '错误数', value: '111' }
       ],
       collapse: true
     },
@@ -291,7 +291,7 @@ const CustomCardNode = () => {
       }
     });
   });
-  
+
 
   return <div ref={container}></div>;
 };