+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - feat: Floyd Warshall shortest path algorithm;
 - feat: built-in arrows;
 - feat: built-in markers;
+- fix: force layout with addItem and relayout.
 
 #### 3.5.7
 - feat: shouldBegin for click-select behavior;